@MichelleArk 2024-05-22
I've done some local 🎩 benchmarking of `dbt clone` against bigquery (since it supports zero-copy cloning) to compare performance on this branch vs main. 

🎩  Setup: 
```
dbt run --target prod
mkdir state
mv target/manifest.json state/
```

### On a project with a chain of 5 models (i.e. 1 -> 2 -> 3 -> 4 -> 5):
Recording average execution time of `dbt clone --state state --full-refresh --threads 5` across 5 runs:
* main: 15.38s, 34% cpu utilization
* this branch: 10.17s, 48% cpu utilization

### On a project with a chain of 10 models:
Recording average execution time of `dbt clone --state state --full-refresh --threads 5` across 5 runs:
* main: 20.41s, 20% cpu utilization
* this branch: 9.02s, 57% cpu utilization